Don't hide information if blips look ugly
This commit changes the behaviour that places the blips on the graph to
not simply hide blips that don't fit nicely without collisions.

As it stands today, blips (and the associated lists) simply don't
display items that couldn't be displayed without collisions.

The assumption is that we rather have a slightly less pleasant graph
than to not display information at all.