[BUGFIX] Travis config adjustments
This adds the libgconf package as an addon
to Travis for getting Cypress to run. Apparently
travis upgraded to Ubuntu 16.4 where this package
is not intrinsically included anymore.

Find more about this at
https://github.com/cypress-io/cypress/issues/4069

Besides, the TEST_URL environment variable is added
as a travis.yml directive in order to work across forks.