Update dependencies to support node 10 & 12
Dependencies rely on specific node ABI versions, so these needed to be
updated in order to support node 10 & 12.

Webpack updated as it has nested dependencies that at least on mac
require older node versions.

These are now compatible with node 10 and 12, meaning the project can be
built with node 10 or above.