Require fonts assets manually to get them included
After upgrade to Webpack 5, our fonts are not included anymore. The
problem seems to be the `css-loader` config that has disabled parsing
of `urls`:

```
    MiniCssExtractPlugin.loader,
    {
        loader: 'css-loader',
        options: { importLoaders: 1, modules: 'global', url: false },
    },
```

The build is not working if we enable url parsing, and I can see that
images already are being required manually:

```
    require('./images/tech-radar-landing-page-wide.png')
```

The `fonts` vs `images` folder is a bit stupid, but it's like that
in old config, as well.

Leaving it like this, until someone finds a better approach.